Fix list_builder
